requirements.txt
opencv-python==4.8.1.78
Pillow==10.0.0
numpy==1.24.3
tqdm==4.66.1
